feat(stock-module): enhance StockSearchPanel and translations for improved user experience
- Added search functionality and hints in both English and Chinese translations to guide users in adding stock favorites.
- Updated StockSearchPanel to streamline the search process and removed unnecessary state management.
- Refactored StockTab to integrate StockSidebar for better organization and usability.
- Adjusted CSS styles for sidebar navigation icons to improve visual consistency.
