+++ i18n/zh-CN.json
@@ -1138,9 +1138,14 @@
       "toggle": "展开/收起侧边栏",
       "favorites": "自选",
       "addFavorite": "添加自选",
+      "search": "搜索",
       "favoritesHint": "管理您的自选股列表",
-      "addFavoriteHint": "搜索股票代码添加到自选",
+      "addFavoriteHint": "输入股票代码并加载行情",
+      "searchHint": "按名称或代码搜索股票",
       "noFavorites": "暂无自选股"
+    },
+    "addFavorite": {
+      "aria": "添加自选股票"
     },
     "empty": {
       "name": "未选择股票",
