+++ src/renderer/components/states/maxExpand/components/stock/components/StockTab.tsx
@@ -19,9 +19,8 @@
  * @author 鸡哥
  */
 
-import { useState, type ReactElement } from 'react';
+import { type ReactElement } from 'react';
 import { useTranslation } from 'react-i18next';
-import { SvgIcon } from '../../../../../../utils/SvgIcon';
 import { useStockMarketData } from '../hooks/useStockMarketData';
 import {
   formatStockChange,
@@ -31,9 +30,7 @@ import {
 } from '../utils/formatters';
 import { StockKlineChart } from './StockKlineChart';
 import { StockMetrics } from './StockMetrics';
-import { StockSearchPanel } from './StockSearchPanel';
-
-type SidebarTab = 'favorites' | 'search';
+import { StockSidebar } from './StockSidebar';
 
 /**
  * 渲染 MaxExpand 股票行情分页。
@@ -43,76 +40,21 @@ export function StockTab(): ReactElement {
   const { t } = useTranslation();
   const market = useStockMarketData();
   const trendClass = getStockTrendClass(market.quote?.changePercent);
-  const [sidebarCollapsed, setSidebarCollapsed] = useState(true);
-  const [sidebarTab, setSidebarTab] = useState<SidebarTab>('favorites');
 
   return (
     <div className="max-expand-settings stock-tab-container">
-      <nav className="stock-sidebar-nav">
-        <button
-          className={`stock-sidebar-nav-btn${sidebarTab === 'favorites' && !sidebarCollapsed ? ' active' : ''}`}
-          type="button"
-          aria-label={t('stockTab.sidebar.favorites')}
-          onClick={() => { setSidebarTab('favorites'); setSidebarCollapsed(false); }}
-        >
-          <img src={SvgIcon.STOCK_CHOOSE} alt="" className="stock-sidebar-nav-icon-upper" />
-        </button>
-        <button
-          className={`stock-sidebar-nav-btn${sidebarTab === 'search' && !sidebarCollapsed ? ' active' : ''}`}
-          type="button"
-          aria-label={t('stockTab.sidebar.addFavorite')}
-          onClick={() => { setSidebarTab('search'); setSidebarCollapsed(false); }}
-        >
-          <img src={SvgIcon.PLUS} alt="" className="stock-sidebar-nav-icon-upper" />
-        </button>
-        <button
-          className="stock-sidebar-nav-btn"
-          type="button"
-          aria-label={t('stockTab.sidebar.toggle')}
-          onClick={() => setSidebarCollapsed(!sidebarCollapsed)}
-        >
-          <img src={sidebarCollapsed ? SvgIcon.EXPAND : SvgIcon.COLLAPSE} alt="" className="stock-sidebar-nav-icon" />
-        </button>
-      </nav>
-
-      <aside className={`stock-tab-sidebar${sidebarCollapsed ? ' collapsed' : ''}`}>
-        {!sidebarCollapsed && (
-          <>
-            {sidebarTab === 'favorites' ? (
-              <div className="stock-favorites-panel">
-                <div className="stock-tab-sidebar-header">
-                  <div>
-                    <div className="stock-tab-title">{t('stockTab.sidebar.favorites')}</div>
-                    <p className="stock-tab-subtitle">{t('stockTab.sidebar.favoritesHint')}</p>
-                  </div>
-                </div>
-                <div className="stock-favorites-empty">{t('stockTab.sidebar.noFavorites')}</div>
-              </div>
-            ) : (
-              <>
-                <div className="stock-tab-sidebar-header">
-                  <div>
-                    <div className="stock-tab-title">{t('stockTab.sidebar.addFavorite')}</div>
-                    <p className="stock-tab-subtitle">{t('stockTab.sidebar.addFavoriteHint')}</p>
-                  </div>
-                </div>
-                <StockSearchPanel
-                  symbol={market.symbol}
-                  period={market.period}
-                  loading={market.loading}
-                  searching={market.searching}
-                  searchResults={market.searchResults}
-                  onSelectSymbol={market.selectSymbol}
-                  onPeriodChange={market.setPeriod}
-                  onRefresh={() => { void market.refresh(); }}
-                  onSearch={market.search}
-                  onClearSearchResults={market.clearSearchResults}
-                />
-              </>
-            )}
-          </>
-        )}
-      </aside>
+      <StockSidebar
+        symbol={market.symbol}
+        period={market.period}
+        loading={market.loading}
+        searching={market.searching}
+        searchResults={market.searchResults}
+        onSelectSymbol={market.selectSymbol}
+        onPeriodChange={market.setPeriod}
+        onRefresh={() => { void market.refresh(); }}
+        onSearch={market.search}
+        onClearSearchResults={market.clearSearchResults}
+      />
 
       <main className="stock-tab-main">
         <div className="stock-tab-main-header">
